requirements.txt
# Discord API
discord.py[voice]

# Google Translate API
googletrans
